fix(agent-service): add download permission check in context_builder
无权限群发图片时 context_builder 无条件调 tool-service 下载，
Lark API 返回错误导致 agent-service → tool-service 100% 错误率告警。
将 check_group_allows_download 提取为共享模块，在图片下载前检查权限，
禁止下载的群跳过未缓存图片的飞书下载流程。
